+++ common/lib/xmodule/xmodule/js/spec/video/video_control_spec.js
@@ -2,15 +2,23 @@
   describe('VideoControl', function() {
     var state, videoControl, oldOTBD;
 
-    function initialize() {
-      loadFixtures('video_all.html');
+    function initialize(fixture) {
+      if (fixture) {
+        loadFixtures(fixture);
+      } else {
+        loadFixtures('video_all.html');
+      }
       state = new Video('#example');
       videoControl = state.videoControl;
     }
 
+    function initializeYouTube() {
+        initialize('video.html');
+    }
+
     beforeEach(function(){
         oldOTBD = window.onTouchBasedDevice;
-        window.onTouchBasedDevice = jasmine.createSpy('onTouchBasedDevice').andReturn(false);
+        window.onTouchBasedDevice = jasmine.createSpy('onTouchBasedDevice').andReturn(null);
     });
 
     afterEach(function() {
@@ -75,13 +83,13 @@
 
       describe('when on a touch based device', function() {
         beforeEach(function() {
-          window.onTouchBasedDevice.andReturn(true);
+          window.onTouchBasedDevice.andReturn(['iPad']);
           initialize();
         });
 
         it('does not add the play class to video control', function() {
-          expect($('.video_control')).not.toHaveClass('play');
-          expect($('.video_control')).not.toHaveAttr('title', 'Play');
+          expect($('.video_control')).toHaveClass('play');
+          expect($('.video_control')).toHaveAttr('title', 'Play');
         });
       });
     });
@@ -146,6 +154,130 @@
           });
         });
       });
+    });
+
+    describe('Play placeholder', function () {
+
+      beforeEach(function () {
+        this.oldYT = window.YT;
+
+        jasmine.stubRequests();
+        window.YT = {
+          Player: function () { },
+          PlayerState: this.oldYT.PlayerState,
+          ready: function (callback) {
+              callback();
+          }
+        };
+
+        spyOn(window.YT, 'Player');
+      });
+
+      afterEach(function () {
+        window.YT = this.oldYT;
+      });
+
+
+      it ('works correctly on calling proper methods', function () {
+        initialize();
+        var btnPlay = state.el.find('.btn-play');
+
+        videoControl.showPlayPlaceholder();
+
+        expect(btnPlay).not.toHaveClass('is-hidden');
+        expect(btnPlay).toHaveAttrs({
+          'aria-hidden': 'false',
+          'tabindex': 0
+        });
+
+        videoControl.hidePlayPlaceholder();
+
+        expect(btnPlay).toHaveClass('is-hidden');
+        expect(btnPlay).toHaveAttrs({
+          'aria-hidden': 'true',
+          'tabindex': -1
+        });
+      });
+
+      var cases = [
+        {
+          name: 'PC',
+          isShown: false,
+          isTouch: null
+        },
+        {
+          name: 'iPad',
+          isShown: true,
+          isTouch: ['iPad']
+        },
+        {
+          name: 'iPhone',
+          isShown: false,
+          isTouch: ['iPhone']
+        }
+      ];
+
+      $.each(cases, function(index, data) {
+        var message = [
+            (data.isShown) ? 'is' : 'is not',
+            ' shown on',
+            data.name
+          ].join('');
+
+        it(message, function () {
+          window.onTouchBasedDevice.andReturn(data.isTouch);
+          initialize();
+          var btnPlay = state.el.find('.btn-play');
+
+          if (data.isShown) {
+            expect(btnPlay).not.toHaveClass('is-hidden');
+          } else {
+            expect(btnPlay).toHaveClass('is-hidden');
+          }
+        });
+      });
+
+      it('is shown on paused video on iPad in HTML5 player', function () {
+        window.onTouchBasedDevice.andReturn(['iPad']);
+        initialize();
+        var btnPlay = state.el.find('.btn-play');
+
+        videoControl.play();
+        videoControl.pause();
+
+        expect(btnPlay).not.toHaveClass('is-hidden');
+      });
+
+      it('is hidden on playing video on iPad in HTML5 player', function () {
+        window.onTouchBasedDevice.andReturn(['iPad']);
+        initialize();
+        var btnPlay = state.el.find('.btn-play');
+
+        videoControl.play();
+
+        expect(btnPlay).toHaveClass('is-hidden');
+      });
+
+      it('is hidden on paused video on iPad in YouTube player', function () {
+        window.onTouchBasedDevice.andReturn(['iPad']);
+        initializeYouTube();
+        var btnPlay = state.el.find('.btn-play');
+
+        videoControl.play();
+        videoControl.pause();
+
+        expect(btnPlay).toHaveClass('is-hidden');
+      });
+
+    });
+
+    it('show', function () {
+      initialize();
+      var controls = state.el.find('.video-controls');
+      controls.addClass('is-hidden');
+
+      videoControl.show();
+      expect(controls).not.toHaveClass('is-hidden');
     });
   });