+++ CHANGELOG.rst
@@ -5,6 +5,14 @@ These are notable changes in edx-platform.  This is a rolling list of changes,
 in roughly chronological order, most recent first.  Add your entries at or near
 the top.  Include a label indicating the component affected.
 
+Blades: Video player improvements:
+  - Disable edX controls for iPhone (native controls are used).
+  - Disable volume and speed controls for iPad.
+  - controls becomes visible after click on video or play placeholder to avoid
+    issues with YouTube API.
+  - Captions becomes visible just after full initialization of video player.
+  - Fix blinking of captions after video player initialization. BLD-206.
+
 LMS: Fix answer distribution download for small courses. LMS-922, LMS-811
 
 Blades: Add template for the zooming image in studio. BLD-206.